added EBNeRD notebook
